real/requirements.txt
# LangGraph 및 AI 관련 (자동 버전 선택)
langgraph
langchain
langchain-openai
langchain-community
openai

# MCP 관련 패키지 추가
langchain-mcp-adapters
python-dotenv

# FastAPI 및 웹 서버
fastapi
uvicorn[standard]
pydantic

# 데이터 처리
pandas
numpy
openpyxl

# 머신러닝 (RAG용)
scikit-learn

# HTTP 클라이언트
requests

# 기타 유틸리티
python-dotenv
typing-extensions